std: dump old std::uv API.. move remaining tests into uv::ll
